DanmakuClient 底层 userId 相关由 int 提升为 long 修复 LiveStartMessage 有时无法正确解析的问题 修复 TarString 可能不正确的反序列化行为

+++ src/Executorlibs.Bilibili.Protocol/Clients/DanmakuClientBase.cs
@@ -383,7 +383,7 @@ namespace Executorlibs.Bilibili.Protocol.Clients
             }
         }
 
-        protected byte[] CreateJoinRoomPayload(int roomId, int userId, string token)
+        protected byte[] CreateJoinRoomPayload(int roomId, long userId, string token)
         {
             byte[] json = JsonSerializer.SerializeToUtf8Bytes(new
             {
